Bigger instr

+++ internal/ast/assign_stmt.go
@@ -30,6 +30,6 @@ func (s *AssignStmt) RunPass(ctx *Context, pass Pass) {
 	switch pass {
 	case Emit:
 		emitter := ctx.Emitter()
-		emitter.Emit(s.Pos(), runtime.OpStore, uint16(len(s.Lvalues)), 0)
+		emitter.Emit(s.Pos(), runtime.OpStore, uint32(len(s.Lvalues)), 0)
 	}
 }
