+++ internal/ast/return_stmt.go
@@ -16,6 +16,6 @@ func (s *ReturnStmt) RunPass(ctx *Context, pass Pass) {
 
 	switch pass {
 	case Emit:
-		ctx.Emitter().Emit(s.Pos(), runtime.OpRet, uint16(len(s.Values)), 0)
+		ctx.Emitter().Emit(s.Pos(), runtime.OpRet, uint32(len(s.Values)), 0)
 	}
 }
